PycharmProjects/case-study-master/accounts/urls.py
from django.urls import path
from django.urls import include
from rest_framework import routers
from .views import UserViewSet


router = routers.DefaultRouter()
router.register('users', UserViewSet)

urlpatterns = [
    path('', include(router.urls)),
]